fix: index by genbank instead of duplicate strain
Since some strains (or isolates) were resequenced resulting in duplicate
strain names in the zika dataset, the original zika build was breaking.

This fixes the Snakefile to index records on GenBank ID instead and swap
in the strain name at the end.

Other changes include updating the list of dropped strain names to
GenBank IDs and updating the example sequence fasta headers.