Add rules for merging USVI data with NCBI GenBank ingested data.
The original Zika build contained USVI data that had been posted publiclly to
GitHub but not yet submitted to NCBI GenBank. This commit adds rules to merge
the USVI data with the NCBI GenBank data.

Since USVI does not have a genbank_accession column, we create a new accession column
for both USVI and NCBI GenBank accessions. This accession column is then used
as the strain_id column for the phylogenetic build.

Since auspice automagically generates a NCBI GenBank url for "genbank_accession" fields,
we use a "url" field instead, allowing for a mix of GenBank and GitHub urls
to be used in the strain popup window.