ingest: Always provide default config values
Part of work to update this repo to match the pathogen-repo-template.

Always provide default config values that can then easily be overridden
with --configfiles/--config options. This makes it simpler to change a
subset of config values or extend the default configs.

Renames the config.yaml to defaults.yaml to reflect this change.

https://github.com/nextstrain/mpox/commit/1e01fedb5fa965124dabb0cafef4aa94b298c44f