Move accession to the first column of metadata_all.tsv
During the merge of Usvi data and GenBank data, the accession field ended
up as the last column. This caused confusion as the first column was named
"genbank_accession" which could be mistaken for the strain ID.

This commit moves the "accession" column to the first column such that
"accession" and "genbank_accession" are next to each other; hopefully,
providing clarity that "accession" is being used as the strain ID, while
"genbank_accession" can be used to generate a url to the NCBI GenBank record
if provided.