ingest/phylo GH Actions: add `trial_name` inputs
Allow trial builds of the ingest and phylogenetic workflows to be
uploaded to trial S3 URLs.

The `trial_name` inputs are _not_ added to the `workflow_call` trigger
since the `workflow_call` trigger is only being used by our automated
ingest-to-phylogenetic workflow. This may change in the future, but
leaving it out for now.

For phylogenetic workflow to run with trial outputs from the ingest
workflow, the phylogenetic workflow needs to accept inputs from the
config. I'll make these changes separately.