Add custom role to read config values
Code entirely by cursor.ai + some manual and prompt-based debugging. Main prompts:

---

I'm in @tutorial.rst and I want to be able to write a custom interpreted text roles to do something like

:configvalue:`phylogenetic/defaults/config.yaml:strain_id_field`

and have it replaced by the corresponding value in `phylogenetic/defaults/config.yaml` of `strain_id_field` which is "accession". The exact syntax used can vary.

This will require writing some custom python code to provide this functionality in sphinx.

---

That's working great. In @tutorial.rst I actually want to reference the yaml field `inputs[0].metadata` however this renderes as `<config value 'inputs[0].metadata' not found in phylogenetic/defaults/config.yaml>`.

---

This is working great for adding the value inline. For some values, e.g.

:configvalue:`ingest/defaults/config.yaml:ncbi_datasets_fields`

The output is a list and would be better rendered as a YAML codeblock by itself rather than inline. Is this possible?

---